requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
datetime-truncate==1.1.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.5
django-bower==5.2.0
django-celery-beat==2.7.0
django-import-export==4.3.4
django-jazzmin==3.0.1
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-timezone-field==7.1
et_xmlfile==2.0.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
mysqlclient==2.2.7
openpyxl==3.1.5
prompt_toolkit==3.0.50
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
